add corrections task19

+++ task19/Program.cs
@@ -2,9 +2,9 @@
 // причем X ≠ 0 и Y ≠ 0
 
 Console.WriteLine("Введите координату X:");
-int X = int.Parse(Console.ReadLine());
+double X = double.Parse(Console.ReadLine());
 Console.WriteLine("Введите координату Y:");
-int Y = int.Parse(Console.ReadLine());
+double Y = double.Parse(Console.ReadLine());
 
 if(X != 0 && Y != 0)
 {
